feat: introduce showStreamDialog and replace showStreamMessageModal
This commit introduces a new generic dialog function `showStreamDialog` and replaces the existing `showStreamMessageModal` with it.

`showStreamDialog` provides a consistent way to display modals with customized transitions and backdrop effects throughout the application. It's a wrapper around `showGeneralDialog` and offers more flexibility in its usage compared to the more specific `showStreamMessageModal`.

Key changes:
- `showStreamMessageModal` has been removed.
- `showStreamDialog` is now used in `MessageWidget` for displaying message actions, moderation actions, and confirmation dialogs.
- `showStreamDialog` ensures that the dialog content is wrapped in a `Center` and `SingleChildScrollView` when `useSafeArea` is true, providing better layout for content that might exceed screen boundaries.
- The new `showStreamDialog` function is exported for external use.
