+++ packages/stream_chat_flutter/lib/stream_chat_flutter.dart
@@ -90,6 +90,7 @@ export 'src/misc/info_tile.dart';
 export 'src/misc/markdown_message.dart';
 export 'src/misc/option_list_tile.dart';
 export 'src/misc/reaction_icon.dart';
+export 'src/misc/stream_modal.dart';
 export 'src/misc/stream_neumorphic_button.dart';
 export 'src/misc/swipeable.dart';
 export 'src/misc/thread_header.dart';
